Moved FTPConfiguration to a new class
